% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simcostar-package.R
\docType{package}
\name{rsimcosta-package}
\alias{rsimcosta}
\alias{rsimcosta-package}
\title{rsimcosta: Client for the SIMCOSTA Buoy Data API}
\description{
Download and cache oceanographic and meteorological buoy data from the SIMCOSTA platform (\url{https://simcosta.furg.br}). Provides incremental downloads with a local SQLite cache and coverage tracking so repeated queries only fetch missing time windows.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/leoniedu/rsimcosta}
  \item Report bugs at \url{https://github.com/leoniedu/rsimcosta/issues}
}

}
\author{
\strong{Maintainer}: Eduardo Leoni \email{e.leoni@gmail.com}

}
\keyword{internal}
